Save invitation token in local storage and accept invite after susccesful user signin
